Move ConservationStatus and related contents to conservation_status module
